requirements.txt
langgraph
streamlit
yfinance
pandas
openai
google-generativeai
requests
python-dotenv
plotly
numpy
networkx
matplotlib
langchain-tavily
